tournament-hub-game-server/requirements.txt
fastapi
uvicorn
requests
cryptography 
pynacl
multiversx-sdk 
websockets
pika